Make modifications on User, Person and SignUpRequest models
