+++ utility/sample_data/user.json
@@ -1,7 +1,5 @@
 [
     {
-        "preferred_name": "Doğukan",
-        "full_name": "Doğukan Teber",
         "email": "dogukan@example.com",
         "username": "dogukan",
         "password": "dogukan",
@@ -9,260 +7,186 @@
         "is_superuser": true
     },
     {
-        "preferred_name": "John",
-        "full_name": "John Doe",
         "email": "john.doe@example.com",
         "username": "johndoe",
         "password": "johndoe123"
     },
     {
-        "preferred_name": "Emily",
-        "full_name": "Emily Smith",
         "email": "emily.smith@example.com",
         "username": "emilysmith",
         "password": "emily123"
     },
     {
-        "preferred_name": "Michael",
-        "full_name": "Michael Johnson",
         "email": "michael.johnson@example.com",
         "username": "michaeljohnson",
         "password": "michael123"
     },
     {
-        "preferred_name": "Sophia",
-        "full_name": "Sophia Garcia",
         "email": "sophia.garcia@example.com",
         "username": "sophiagarcia",
         "password": "sophia123"
     },
     {
-        "preferred_name": "William",
-        "full_name": "William Anderson",
         "email": "william.anderson@example.com",
         "username": "williamanderson",
         "password": "william123"
     },
     {
-        "preferred_name": "Olivia",
-        "full_name": "Olivia Brown",
         "email": "olivia.brown@example.com",
         "username": "oliviabrown",
         "password": "olivia123"
     },
     {
-        "preferred_name": "James",
-        "full_name": "James Wilson",
         "email": "james.wilson@example.com",
         "username": "jameswilson",
         "password": "james123"
     },
     {
-        "preferred_name": "Emma",
-        "full_name": "Emma Lee",
         "email": "emma.lee@example.com",
         "username": "emmalee",
         "password": "emma123"
     },
     {
-        "preferred_name": "David",
-        "full_name": "David Nguyen",
         "email": "david.nguyen@example.com",
         "username": "davidnguyen",
         "password": "david123"
     },
     {
-        "preferred_name": "Sarah",
-        "full_name": "Sarah Graham",
         "email": "sarah.graham@example.com",
         "username": "sarahgraham",
         "password": "sarah123"
     },
     {
-        "preferred_name": "Kevin",
-        "full_name": "Kevin Miller",
         "email": "kevin.miller@example.com",
         "username": "kevinmiller",
         "password": "kevin123"
     },
     {
-        "preferred_name": "Jessica",
-        "full_name": "Jessica Lopez",
         "email": "jessica.lopez@example.com",
         "username": "jessicalopez",
         "password": "jessica123"
     },
     {
-        "preferred_name": "Alex",
-        "full_name": "Alex Robinson",
         "email": "alex.robinson@example.com",
         "username": "alexrobinson",
         "password": "alex123"
     },
     {
-        "preferred_name": "Amanda",
-        "full_name": "Amanda Baker",
         "email": "amanda.baker@example.com",
         "username": "amandabaker",
         "password": "amanda123"
     },
     {
-        "preferred_name": "Mark",
-        "full_name": "Mark Taylor",
         "email": "mark.taylor@example.com",
         "username": "marktaylor",
         "password": "mark123"
     },
     {
-        "preferred_name": "Grace",
-        "full_name": "Grace Hernandez",
         "email": "grace.hernandez@example.com",
         "username": "gracehernandez",
         "password": "grace123"
     },
     {
-        "preferred_name": "Robert",
-        "full_name": "Robert Allen",
         "email": "robert.allen@example.com",
         "username": "robertallen",
         "password": "robert123"
     },
     {
-        "preferred_name": "Rachel",
-        "full_name": "Rachel Thomas",
         "email": "rachel.thomas@example.com",
         "username": "rachelthomas",
         "password": "rachel123"
     },
     {
-        "preferred_name": "Daniel",
-        "full_name": "Daniel Wilson",
         "email": "daniel.wilson@example.com",
         "username": "danielwilson",
         "password": "daniel123"
     },
     {
-        "preferred_name": "Michelle",
-        "full_name": "Michelle Carter",
         "email": "michelle.carter@example.com",
         "username": "michellecarter",
         "password": "michelle123"
     },
     {
-        "preferred_name": "Andrew",
-        "full_name": "Andrew Wong",
         "email": "andrew.wong@example.com",
         "username": "andrewwong",
         "password": "andrew123"
     },
     {
-        "preferred_name": "Natalie",
-        "full_name": "Natalie Parker",
         "email": "natalie.parker@example.com",
         "username": "natalieparker",
         "password": "natalie123"
     },
     {
-        "preferred_name": "Jason",
-        "full_name": "Jason Miller",
         "email": "jason.miller@example.com",
         "username": "jasonmiller",
         "password": "jason123"
     },
     {
-        "preferred_name": "Samantha",
-        "full_name": "Samantha White",
         "email": "samantha.white@example.com",
         "username": "samanthawhite",
         "password": "samantha123"
     },
     {
-        "preferred_name": "Christopher",
-        "full_name": "Christopher Lee",
         "email": "christopher.lee@example.com",
         "username": "christopherlee",
         "password": "christopher123"
     },
     {
-        "preferred_name": "Melissa",
-        "full_name": "Melissa Chen",
         "email": "melissa.chen@example.com",
         "username": "melissachen",
         "password": "melissa123"
     },
     {
-        "preferred_name": "Jonathan",
-        "full_name": "Jonathan Murphy",
         "email": "jonathan.murphy@example.com",
         "username": "jonathanmurphy",
         "password": "jonathan123"
     },
     {
-        "preferred_name": "Stephanie",
-        "full_name": "Stephanie Allen",
         "email": "stephanie.allen@example.com",
         "username": "stephanieallen",
         "password": "stephanie123"
     },
     {
-        "preferred_name": "Brandon",
-        "full_name": "Brandon Perez",
         "email": "brandon.perez@example.com",
         "username": "brandonperez",
         "password": "brandon123"
     },
     {
-        "preferred_name": "Amy",
-        "full_name": "Amy Turner",
         "email": "amy.turner@example.com",
         "username": "amyturner",
         "password": "amy123"
     },
     {
-        "preferred_name": "Ryan",
-        "full_name": "Ryan Flores",
         "email": "ryan.flores@example.com",
         "username": "ryanflores",
         "password": "ryan123"
     },
     {
-        "preferred_name": "Lauren",
-        "full_name": "Lauren Barnes",
         "email": "lauren.barnes@example.com",
         "username": "laurenbarnes",
         "password": "lauren123"
     },
     {
-        "preferred_name": "Eric",
-        "full_name": "Eric Rivera",
         "email": "eric.rivera@example.com",
         "username": "ericrivera",
         "password": "eric123"
     },
     {
-        "preferred_name": "Hannah",
-        "full_name": "Hannah Cooper",
         "email": "hannah.cooper@example.com",
         "username": "hannahcooper",
         "password": "hannah123"
     },
     {
-        "preferred_name": "Patrick",
-        "full_name": "Patrick Young",
         "email": "patrick.young@example.com",
         "username": "patrickyoung",
         "password": "patrick123"
     },
     {
-        "preferred_name": "Olivia",
-        "full_name": "Olivia Hill",
         "email": "olivia.hill@example.com",
         "username": "oliviahill",
         "password": "olivia123"
     },
     {
-        "preferred_name": "Matthew",
-        "full_name": "Matthew Martin",
         "email": "matthew.martin@example.com",
         "username": "matthewmartin",
         "password": "matthew123"
